Added thriller.js | unedited sumobot.js
