+++ code_example/thriller.js
@@ -1,0 +1,134 @@
+// =======================
+// Sumobot Jr demo program
+// =======================
+
+var five = require("johnny-five");
+var keypress = require('keypress');
+var promise = require('bluebird');
+
+keypress(process.stdin);
+
+var board = new five.Board();
+
+board.on("ready", function() {
+
+  console.log("Welcome to Sumobot Jr!")
+  console.log("Control the bot with the arrow keys, and SPACE to stop.")
+
+  var left_wheel  = new five.Servo({ pin:  9, type: 'continuous' }).stop();
+  var right_wheel = new five.Servo({ pin: 10, type: 'continuous'  }).stop();
+
+
+  process.stdin.resume(); 
+  process.stdin.setEncoding('utf8'); 
+  process.stdin.setRawMode(true); 
+
+  var okay = promise.resolve();
+
+  var beat = 508;
+
+//helper functions  
+  
+function stop() {
+
+      console.log('Stopping');
+      left_wheel.stop();
+      right_wheel.stop();
+}
+
+function turnleft() {
+
+      console.log('Left');
+      left_wheel.ccw();
+      right_wheel.ccw();  
+}
+
+function turnright() {
+
+      console.log('Right');
+      left_wheel.cw();
+      right_wheel.cw();
+}
+
+function forward() {
+
+      console.log('Forward');
+      left_wheel.ccw();
+      right_wheel.cw();
+}
+
+function back() {
+
+      console.log('Backward');
+      left_wheel.cw();
+      right_wheel.ccw(); 
+}
+
+
+  process.stdin.on('keypress', function (ch, key) {
+    
+
+    if ( !key ) return;
+
+
+    if ( key.name == 'q' ) {
+
+      console.log('Quitting');
+      process.exit();
+
+    }
+
+    /* else if ( key.name == 'up' ) {
+
+      console.log('Forward');
+      left_wheel.ccw();
+      right_wheel.cw();
+
+    } else if ( key.name == 'down' ) {
+
+      console.log('Backward');
+      left_wheel.cw();
+      right_wheel.ccw();      
+
+    } else if ( key.name == 'left' ) {
+
+      console.log('Left');
+      left_wheel.ccw();
+      right_wheel.ccw();      
+
+
+    } else if ( key.name == 'right' ) {
+
+      console.log('Right');
+      left_wheel.cw();
+      right_wheel.cw();
+      */
+
+   if ( key.name == 'space' ) {
+    
+        for(var i = 0; i < 10; i++) {
+          if(i == 6){
+            okay = okay.delay(beat * 2);
+            for(var j=0; j < 16; j++){
+            okay = okay.then(forward);
+            okay = okay.delay(beat); 
+            okay = okay.then(back);
+            okay = okay.then(beat);
+      }
+        continue;
+  }
+
+          okay = okay.then(turnleft);
+          okay = okay.delay(4 * beat);
+          okay = okay.then(turnright);
+          okay = okay.delay(4 * beat);
+          okay = okay.then(forward);
+          okay = okay.delay(4 * beat);
+          okay = okay.then(back);
+          okay = okay.delay(4 * beat);
+          okay = okay.then(stop);
+      }
+    }
+  });
+});
+
